```
goos: darwin
goarch: arm64
pkg: github.com/cristalhq/bson
                      │   old.txt   │               new.txt               │
                      │   sec/op    │   sec/op     vs base                │
_cristalhq_Marshal-10   1.735µ ± 3%   1.285µ ± 0%  -25.94% (p=0.000 n=10)

                      │   old.txt    │               new.txt                │
                      │     B/op     │     B/op      vs base                │
_cristalhq_Marshal-10   3.191Ki ± 0%   1.625Ki ± 0%  -49.08% (p=0.000 n=10)

                      │  old.txt   │              new.txt               │
                      │ allocs/op  │ allocs/op   vs base                │
_cristalhq_Marshal-10   36.00 ± 0%   25.00 ± 0%  -30.56% (p=0.000 n=10)
```